requirements.txt
torch
pandas
nltk
numpy
matplotlib
beautifulsoup4
gensim
py-rouge
tqdm
